+++ lib/global_id/identification.rb
@@ -9,9 +9,17 @@ class GlobalID
     end
     alias to_gid to_global_id
 
+    def to_gid_param
+      to_global_id.to_param
+    end
+
     def to_signed_global_id(options = {})
       SignedGlobalID.create(self, options)
     end
     alias to_sgid to_signed_global_id
+
+    def to_sgid_param(options = {})
+      to_signed_global_id(options).to_param
+    end
   end
 end