+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Added #to_gid_param and #to_sgid_param as short-cuts for to_gid.to_param and to_sgid.to_param for convenience.
+
+    *DHH*
+
 *   Added support for Locator.locate_many and Locator.locate_many_signed to efficiently fetch many models.
 
     *Tom Ward, DHH*